Changed the build script to have a "makeRelease" task to build both the plain and module versions of the Javascript and copy it to the release directory.
I took the version number off the .mjs file because presumably if people point to it, they want the latest one, and it will be pain if the version number gets bumped and there ends up with 1.0 and 1.1 and 1.2 etc in that directory.  The version can be taken care of elsewhere (e.g. in a package.json file).